requirements.txt
Django==4.2.3
python-telegram-bot==13.15
environs==9.5.0
requests==2.31.0
